git-svn-id: http://svn.code.sf.net/p/shopizer/code/shopizer-portlet@2876 e80ee2c7-76f3-4005-aeaf-98396743f455

+++ sm-core/src/main/java/com/salesmanager/core/business/catalog/product/model/review/ProductReview.java
@@ -42,7 +42,7 @@ public class ProductReview extends SalesManagerEntity<Long, ProductReview> imple
 	private AuditSection audit = new AuditSection();
 	
 	@Column(name = "REVIEWS_RATING")
-	private Integer reviewRating;
+	private Double reviewRating;
 	
 	@Column(name = "REVIEWS_READ")
 	private Long reviewRead;
@@ -72,11 +72,11 @@ public class ProductReview extends SalesManagerEntity<Long, ProductReview> imple
 		this.id = id;
 	}
 
-	public Integer getReviewRating() {
+	public Double getReviewRating() {
 		return reviewRating;
 	}
 
-	public void setReviewRating(Integer reviewRating) {
+	public void setReviewRating(Double reviewRating) {
 		this.reviewRating = reviewRating;
 	}
